% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{AREA}
\alias{AREA}
\title{Whether or not study sub-areas were used in a given year.}
\format{A data frame with 33 rows and 10 variables
\describe{
  \item{Year}{Study year (1984 - 2016)}
  \item{A}{Logical. Was sub-area A used in a given year.}
  \item{B}{Logical. Was sub-area B used in a given year.}
  \item{C}{Logical. Was sub-area C used in a given year.}
  \item{D}{Logical. Was sub-area D used in a given year.}
  \item{E}{Logical. Was sub-area E used in a given year.}
  \item{F}{Logical. Was sub-area F used in a given year.}
  \item{OBK}{Logical. Was sub-area OBK used in a given year.}
  \item{NBK}{Logical. Was sub-area NBK used in a given year.}
  \item{TBK}{Logical. Was sub-area TBK used in a given year.}
  }}
\description{
Data frame showing whether a study sub area was used in a given year.
}
