% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{FLEDGE_DATA}
\alias{FLEDGE_DATA}
\title{Number of fledglings recorded in each territory 1984 - 2016.}
\format{A data frame with 3,754 rows and 7 variables
\describe{
  \item{Status}{Status of territory. Either 'leapfrog' (i.e. inland) or 'resident' (i.e. coastal).}
  \item{NestID}{Whether a nest was found in the territory.}
  \item{Sub_Area}{Study sub-area where territory is located.}
  \item{Year}{Year when fledgling numbers were recorded.}
  \item{Fledge_ChickInfo}{Fledgling numbers estimated from individual chick records.}
  \item{Fledge_NestInfo}{Fledgling numbers estimated from nest records.}
  \item{Fledge_NestCheck}{Fledgling numbers estimated from individual nest cheks.}
  }}
\description{
Number of fledglings recorded in each territory 1984 - 2016.
}
