% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{GRID_DENSITY}
\alias{GRID_DENSITY}
\title{Territory density in 50m x 50m grid squares between 1984 - 2016.}
\format{A data frame with 10,303 rows and 3 variables
\describe{
  \item{ID}{Unique ID for each grid square.}
  \item{DENSITY}{Number of territory centre points in the grid square.}
  \item{Year}{Year of density measurement.}
  }}
\description{
Territory density in 50m x 50m grid squares between 1984 - 2016.
}
