% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{GULLY}
\alias{GULLY}
\title{SpatialPolygonsDataFrame of tidal gullies.}
\format{A SpatialPolygonsDataFrame with 5 features (gullies).
\describe{
  }}
\description{
SpatialPolygonsDataFrame of tidal gullies.
}
