% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Nest_elevation}
\alias{Nest_elevation}
\title{Elevation of H ostralegus nests 1995 - 2016.}
\format{A data frame with 3,221 rows and 6 variables.
\describe{
  \item{Year}{Year of measurement.}
  \item{Z}{Nest elevation (cm).}
  \item{Method}{Measurement method. Either in-situ (DGPS) or ex-situ (LID).}
  \item{MaleID}{Unique identifier for male parents of the nest.}
  \item{Area}{Study sub-area where nest is found.}
  \item{Status}{Status of territory. Either 'leapfrog' (i.e. inland) or 'resident' (i.e. coastal).}
  }}
\description{
Elevation of H ostralegus nests 1995 - 2016.
}
