% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figS7.R
\name{plot_figS7}
\alias{plot_figS7}
\title{Make grid plot of 2013 - 2016 settlement}
\usage{
plot_figS7(palette = "Greens")
}
\arguments{
\item{palette}{Colour-brewer palette to use.}
}
\value{
ggplot object.
}
\description{
Plot showin settlement of H. ostralegus between
2013 and 2016 overlayed on a) fledglin output,
b) conspecific density, c) elevation.
}
