% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Raw_fledge_data}
\alias{Raw_fledge_data}
\title{Total number of fledglings produced in 50m x 50m grid squares each year.}
\format{A data frame with 3,754 rows and 3 variables.
\describe{
  \item{Year}{Year of measurement.}
  \item{POLY_ID}{Unique ID for each grid square.}
  \item{Fledge_est}{Total number of fledglings in each grid square each year.}
  }}
\description{
Total number of fledglings produced in 50m x 50m grid squares each year.
}
