% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{SETTLE_poisson_100}
\alias{SETTLE_poisson_100}
\title{Territory settlement of H ostralegus in 100m x 100m grid squares 1985 - 2016.}
\format{A data frame with 2,535 rows and 11 variables.
\describe{
  \item{Year}{Year of measurement.}
  \item{POLY_ID}{Unique ID for each grid square.}
  \item{nr_Settle}{Total number of settlements occuring within grid square in given year.}
  \item{lgl_Settle}{Have any settlements occurred within grid square in a given year.}
  \item{Median2}{Median elevation of grid square (cm).}
  \item{Sub_Area}{Study sub-area of grid square.}
  \item{Gully_Dist}{Distance to nearest tidal gully.}
  \item{Coast_Dist}{Distance to the coastline.}
  \item{Grid_area}{Area of grid square after removing unusable areas (elevation <20cm above MHT in 1971)}
  \item{rel_density3yr}{Relative density of H. ostralegus territories in the past 3 years before settlement.}
  \item{rel_fledge_est3yr}{Relative number of fledglings in the past 3 years before settlement.}
  }}
\description{
Territory settlement of H ostralegus in 100m x 100m grid squares 1985 - 2016.
}
