% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{TERR_PTS}
\alias{TERR_PTS}
\title{SpatialPointsDataFrame of territory centre points 1984 - 2016.}
\format{A data frame with 3,754 rows and 7 variables.
\describe{
  \item{MaleCode}{Male colour band code.}
  \item{MaleID}{Unique male ID.}
  \item{Status}{Status of territory. Either 'leapfrog' (i.e. inland) or 'resident' (i.e. coastal).}
  \item{Median}{Median elevation of grid square (cm).}
  \item{NestID}{Whether a nest was found in the territory.}
  \item{Gully_Dist}{Distance to nearest tidal gully.}
  \item{Coast_Dist}{Distance to the coastline.}
  \item{Sub_Area}{Study sub-area of grid square.}
  \item{Year}{Year of measurement.}
  }}
\description{
SpatialPointsDataFrame of territory centre points 1984 - 2016.
}
