% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{VacatedTerritories}
\alias{VacatedTerritories}
\title{Territory vacancy patterns of H. ostralegus 1984 - 2015}
\format{A data frame with 3,116 rows and 7 variables
\describe{
  \item{Median2}{Median elevation of grid square (cm above MHT in 1971).}
  \item{Year}{Year of measurement.}
  \item{Coast_Dist}{Distance to the coastline.}
  \item{Gully_Dist}{Distance to nearest tidal gully.}
  \item{Status}{Status of territory. Either 'leapfrog' (i.e. inland) or 'resident' (i.e. coastal).}
  \item{Sub_Area}{Study sub-area of grid square.}
  \item{Vacated}{Whether or not a territory was vacated the following year.}
  }}
\description{
Data frame showing territory vacancy of H. ostralegus on Schiermonnikoog from 1984 - 2015.
Data frame is generated in the vignette 'prepare_data'.
}
